+++ examples/image__unimplemented.rs
@@ -34,13 +34,13 @@ fn main() {
     // let mut result_buffer: Buffer<f32> = Buffer::with_vec(&dims, &ocl_pq.queue());
 
     // // Create a kernel with three arguments corresponding to those in the kernel:
-    // let kernel = ocl_pq.create_kernel("multiply_by_scalar", dims.work_dims())
+    // let kernel = ocl_pq.create_kernel_with_dims("multiply_by_scalar", dims.clone())
     //  .arg_buf(&source_buffer)
     //  .arg_buf(&mut result_buffer)
     // ;
 
     // // Enqueue kernel depending on and creating no events:
-    // kernel.enqueue(None, None);
+    // kernel.enqueue();
 
     // // Read results from the device into the buffer's vector:
     // result_buffer.fill_vec();